refactor: Move getXPContentImage to imageUtils
